requirements.txt
# algos requirements
fastai==1.0.61
librosa==0.8.0
Pillow==7.2.0
pytest==6.0.1
PyYAML==5.3.1
torch==1.4.0
torchvision==0.5.0

# server requirements
connexion[swagger-ui]
python_dateutil
setuptools >= 21.0.0
python-dotenv
azure-storage-blob
azure-cosmos